### 2020-05-22 #2219 @mdmurbach
👋 @arfon Thanks for all your work!

I know we're still waiting on an editor, but I've looked through the list above and listed a few reviewers for now 😄 I'll also plan on fixing the DOI issue flagged above. Please let me know if there is anything else I can be of help with. Hope you're staying healthy; have a good weekend! 🚀

#### Potential reviewers
- `EricaEgg` <-- we've worked in the same lab together, but not on this software
- `amandadumi`
- `shivupa`
- `untzag`
