### 2020-05-22 653cc15
(paper) update the citation for scipy
