+++ joss-paper/paper.md
@@ -62,7 +62,7 @@ Once you've collected your data, the last thing you want to do is worry about ho
 Ensuring that the system under measurement is linear, stable, and causal is an important, but often overlooked part of impedance analysis. `impedance.py` provides several methods  (measurement models, Lin-KK algorithm [@schonleber_method_2014]) for data validation as a part of the same easy-to-use package.
 
 ## Equivalent circuit fitting
-`impedance.py` equivalent circuit fitting combines an extremely flexible circuit definition scheme with a range of available circuit elements -- from simple L, R, C, and constant-phase elements to electrochemistry specific elements such as Warburg, Gerischer [@boukamp_gerischer_element_2003], and transmission-line models [@paasch_theory_1993]. Custom elements can be easily added to a single users installation or to the overall project by [creating an issue](https://github.com/ECSHackWeek/impedance.py/issues/new?assignees=&labels=&template=equivalent-circuit-element-request.md&title=%5BElement%5D). Additional model fitting features include parameter bounds, holding circuit elements constant, weighted fitting, and even saving/loading models to a human readable .json file. Fitting is performed by non-linear least squares regression of the circuit model to impedance data via `curve_fit` from the `scipy.optimize` package [@jones_scipy_2001].
+`impedance.py` equivalent circuit fitting combines an extremely flexible circuit definition scheme with a range of available circuit elements -- from simple L, R, C, and constant-phase elements to electrochemistry specific elements such as Warburg, Gerischer [@boukamp_gerischer_element_2003], and transmission-line models [@paasch_theory_1993]. Custom elements can be easily added to a single users installation or to the overall project by [creating an issue](https://github.com/ECSHackWeek/impedance.py/issues/new?assignees=&labels=&template=equivalent-circuit-element-request.md&title=%5BElement%5D). Additional model fitting features include parameter bounds, holding circuit elements constant, weighted fitting, and even saving/loading models to a human readable .json file. Fitting is performed by non-linear least squares regression of the circuit model to impedance data via `curve_fit` from the `scipy.optimize` package [@2020SciPy-NMeth].
 
 ## Model selection
 One of the biggest challenges in extracting parameters from EIS spectra is quantitatively evaluating the fit of different models. `impedance.py` returns estimated parameter error bars with every fit and comes with a built in method for calculating the confidence bounds of a fit model.
