### 2020-05-22 #2219 @mdmurbach
Fixed the scipy reference issue above 🎉
