### 2020-05-22 #2219 @arfon
@mdmurbach - DOIs should be listed as e.g. `10.1038/s41592-019-0686-2` in the BibTeX please.
