### 2020-05-22 cf6e755
fixes incorrect doi format, related to openjournals/joss-reviews#2219
