### 2020-06-04 f1caf7c
Replaced exampleDataBioLogic.mpr with the same data that's in the .mpt file. Added a .txt version of the same file for readability.
