+++ data/exampleDataBioLogic.txt
@@ -1,0 +1,44 @@
+freq/Hz	Re(Z)/Ohm	Im(Z)/Ohm	
+1.0003201E+003	6.5470886E+001	-3.8998979E-001
+7.7024658E+002	6.3611004E+001	-1.3082615E+000
+5.9291284E+002	6.3786083E+001	4.9220982E-001
+4.5631409E+002	6.6016418E+001	-1.1641068E+000
+3.5101083E+002	6.4581978E+001	1.2355437E+000
+2.7043265E+002	6.2677181E+001	-2.4978859E+000
+2.0793719E+002	6.3067787E+001	1.2588621E+000
+1.6025639E+002	6.4594604E+001	-1.8496882E+000
+1.2330331E+002	6.4560471E+001	3.1567256E+000
+9.4908844E+001	6.3480858E+001	-1.6610308E+000
+7.3059525E+001	6.2966324E+001	-5.2031708E-001
+5.6241814E+001	5.7859802E+001	-9.9445152E-001
+4.3295860E+001	6.3966675E+001	-1.3191714E+000
+3.3346024E+001	6.5947289E+001	-9.2110091E-001
+2.5665239E+001	6.3702545E+001	-1.7817245E+000
+1.9753477E+001	6.5081535E+001	-2.1123164E+000
+1.5199414E+001	6.3539898E+001	-4.5586529E+000
+1.1695360E+001	6.3876205E+001	-1.0484103E+000
+9.0005779E+000	6.4856674E+001	-2.2519822E+000
+6.9382758E+000	6.6270004E+001	-4.8155041E+000
+5.3364062E+000	6.0613422E+001	-7.0270448E+000
+4.1118407E+000	6.5389824E+001	-5.1409044E+000
+3.1603959E+000	6.5957474E+001	-6.4717007E+000
+2.4353178E+000	6.6105698E+001	-8.3983793E+000
+1.8735006E+000	6.7790421E+001	-9.3406420E+000
+1.4435508E+000	6.8793274E+001	-1.2584828E+001
+1.1097296E+000	7.0818695E+001	-1.4015121E+001
+8.5569519E-001	7.3593681E+001	-1.6304792E+001
+6.5806097E-001	7.7258682E+001	-1.8608200E+001
+5.0638443E-001	8.1968651E+001	-2.0657204E+001
+3.8965085E-001	8.6686745E+001	-2.1498911E+001
+2.9992697E-001	9.2175919E+001	-2.1338537E+001
+2.3078397E-001	9.6659569E+001	-1.9389862E+001
+1.7773455E-001	1.0135281E+002	-1.7685032E+001
+1.3685971E-001	1.0487324E+002	-1.4734948E+001
+1.0535786E-001	1.0643014E+002	-1.2445711E+001
+8.1177264E-002	1.0825275E+002	-9.8288803E+000
+6.2444042E-002	1.0905769E+002	-7.8290782E+000
+4.8106518E-002	1.0965545E+002	-5.8287849E+000
+3.7039053E-002	1.1001928E+002	-4.7268991E+000
+2.8483247E-002	1.1061794E+002	-3.7406499E+000
+2.1945218E-002	1.1078029E+002	-2.8922150E+000
+1.6895540E-002	1.1097003E+002	-2.3458567E+000
