### 2020-06-11 2b7b015
Update preprocessing.py
delete extra spaces
