### 2020-06-11 9627cae
added test for readBioLogic if missing freq column added biologic data file missing freq column
