### 2020-06-16 #2349 @whedon
[REVIEW]: impedance.py: A Python package for electrochemical impedance analysis
**Submitting author:** @mdmurbach (<a href="http://orcid.org/0000-0002-6583-5995">Matthew D. Murbach</a>)
**Repository:** <a href="https://github.com/ECSHackWeek/impedance.py" target ="_blank">https://github.com/ECSHackWeek/impedance.py</a>
**Version:** v1.0.1
