### 2020-06-18 #2349 @EricaEgg
@mbobra, I have a question regarding the COI guidelines. One of the authors (Brian Gerwe) and I are both PhD students within UW ChemE, therefore "employed at the same institution". However, we are NOT in the same lab group, so is that acceptable? Just want to make sure we are within the regulations. Thanks!
