### 2020-06-23 #2349 @EricaEgg
> > @mbobra, I have a question regarding the COI guidelines. One of the authors (Brian Gerwe) and I are both PhD students within UW ChemE, therefore "employed at the same institution". However, we are NOT in the same lab group, so is that acceptable? Just want to make sure we are within the regulations. Thanks!
>
> @EricaEgg - do you believe this will affect your ability to give an impartial review? If not, please proceed :-)

Thanks @arfon, it will not. I am happy to review.
