### 2020-07-09 a9cde54
add FAQ file with fitting method
