### 2020-07-13 #2349 @ma-sadeghi
Hi @mbobra, I have a question for you: does the writing have to be strictly formal, or a few informalities here and there are acceptable?
