### 2020-07-13 #2349 @mdmurbach
Hi @ma-sadeghi, thanks for your time with the review!

Yes, you're right. If you look at the timeline/content of the commits ([1](https://github.com/ECSHackWeek/impedance.py/commits?author=Jbonezzi22), [2](https://github.com/ECSHackWeek/impedance.py/commits?author=dacb), [3](https://github.com/ECSHackWeek/impedance.py/commits?author=pangq2)) for the 3 folks who are not listed as authors, they were in the initial setup of the repository structure during the Electrochemical Society (ECS) Hackweek and not in the actual content of the package being published. In this case, I don't think it makes sense to use LOC as a metric for contributions and the four authors were selected as they have all made substantial contributions to the work.
