### 2020-07-17 #2349 @EricaEgg
@mdmurbach The README references an examples/ directory, but I do not see that on the master branch. Should users just reference the readthedocs examples instead? Just want to clarify. Thanks!
