### 2020-07-17 #2349 @mdmurbach
Good catch @EricaEgg! Yes, the directory was moved into docs/source, so you can find them [here](https://github.com/ECSHackWeek/impedance.py/tree/master/docs/source/examples). These notebook files become the examples rendered in the documentation using [nbsphinx](https://nbsphinx.readthedocs.io/en/0.7.1/).

I will update the README to fix that discrepancy as well.
