### 2020-07-17 #2349 @EricaEgg
> Great work everyone! I just have a few documentation recommendations, but other than that I would say it is good to go!
>
> Doc recs:
>
> 1. In the notebook, _fitting_example.ipynb_, part 4a, numpy is called without an import. Import line needs to be added in this block or earlier.
> 2. In the notebook _plotting_example.ipynb_, part 3a, I am getting an error. There could be a bug in the function `circuit.plot`. (error attached as photo)
>
> <img alt="Screen Shot 2020-07-17 at 2 30 21 PM" width="507" src="https://user-images.githubusercontent.com/46461002/87832095-0d004000-c83a-11ea-9e52-f70a309ca514.png">
>
> If the above errors are due to errors I've done, might be useful to mention in documentation for users. Since these are small things, I went ahead and marked the corresponding boxes.

Created an issue on target repo
