### 2020-07-18 2797d5f
fix imports and reference to examples directory, openjournals/joss-reviews#2349
