### 2020-07-18 #2349 @mdmurbach
Awesome, thanks for taking the time to review @EricaEgg! I've added those changes as well as a few thoughts to the plotting error issue (I think it's potentially an adblocker or versioning issue rather than anything related to the package).
