### 2020-07-20 #2349 @EricaEgg
> Awesome, thanks for taking the time to review @EricaEgg! I've added those changes as well as a few thoughts to the plotting error issue (I think it's potentially an adblocker or versioning issue rather than anything related to the package).

Thanks again @mdmurbach! Everything looks great from my end now. Good to go!
