### 2020-07-21 #2349 @mbobra
Thank you for all your hard work @EricaEgg and @ma-sadeghi! I really appreciate it.

@mdmurbach We're almost there! Can you please cut a new release (it doesn't have to be a major release, something like v1.0.1 is fine), [archive your release on Zenodo](https://guides.github.com/activities/citable-code/) to obtain a DOI, and then put that DOI in your `README.md` file? After that I think we're done 🎉
