### 2020-07-21 b8a6138
Adds Zenodo DOI badge
openjournals/joss-reviews#2349
