### 2020-07-21 #2349 @mdmurbach
Linked with Zenodo, released v1.0.1, and added our shiny new DOI to [the README](https://github.com/ECSHackWeek/impedance.py/blob/master/README.md). Thanks for all of your time as well @mbobra! Please let me know if there is anything else you need
