### 2020-08-03 #2349 @danielskatz
@mdmurbach - I notice the zenodo software author list has more people than the JOSS paper author list - is this intentional?  normally, we expect them to be the same, but this is not required.
