### 2020-08-03 #2349 @danielskatz
@mdmurbach - in the paper:
- The journal for Hunter should be Computing in Science & Engineering (the & is missing)
- for Rodrigues et al. Should A.C. be in caps in the title?
