### 2020-08-03 #2349 @mdmurbach
@danielskatz great catch on these. Thanks! 😄

- I've added the `&` to the Hunter paper
- a.c. is not capitalized in the original [Rodrigues reference](https://www.sciencedirect.com/science/article/abs/pii/S0378775399003511?via%3Dihub), so I left that as is
- added the countries to the author list
- Fixed the Zenodo affiliations to match
