### 2020-08-04 #2349 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
